+++ Runtime/Audio/AudioManager.cs
@@ -4,25 +4,24 @@ using UnityEngine.Audio;
 
 namespace F8Framework.Core
 {
-    [UpdateRefresh]
     public class AudioManager : ModuleSingletonMono<AudioManager>, IModule
     {
         /*----------背景音乐----------*/
-        private AudioMusic _audioMusic;
+        public AudioMusic AudioMusic;
         private float _volumeMusic = 1f;
         private bool _switchMusic = true;
         
         /*----------人声----------*/
-        private AudioMusic _audioMusicVoice;
+        public AudioMusic AudioMusicVoice;
         private float _volumeVoice = 1f;
         private bool _switchVoice = true;
         
         /*----------特效声----------*/
-        private AudioMusic _audioMusicBtnClick;
+        public AudioMusic AudioMusicBtnClick;
         
-        private AudioMusic _audioMusicUISound;
+        public AudioMusic AudioMusicUISound;
         
-        private AudioMusic _audioMusicAudioEffect;
+        public AudioMusic AudioMusicAudioEffect;
 
         /*----------一次性特效声----------*/
         private AudioEffect _audioMusicAudioEffect3D;
@@ -46,38 +45,38 @@ namespace F8Framework.Core
             _transform = this.transform;
             GameObject gameObjectMusic = new GameObject("Music", typeof(AudioSource));
             gameObjectMusic.transform.SetParent(_transform);
-            _audioMusic = new AudioMusic();
-            _audioMusic.MusicSource = gameObjectMusic.GetComponent<AudioSource>();
-            _audioMusic.MusicSource.playOnAwake = false;
-            _audioMusic.MusicSource.loop = false;
+            AudioMusic = new AudioMusic();
+            AudioMusic.MusicSource = gameObjectMusic.GetComponent<AudioSource>();
+            AudioMusic.MusicSource.playOnAwake = false;
+            AudioMusic.MusicSource.loop = false;
 
             GameObject gameObjectVoice = new GameObject("Voice", typeof(AudioSource));
             gameObjectVoice.transform.SetParent(_transform);
-            _audioMusicVoice = new AudioMusic();
-            _audioMusicVoice.MusicSource = gameObjectVoice.GetComponent<AudioSource>();
-            _audioMusicVoice.MusicSource.playOnAwake = false;
-            _audioMusicVoice.MusicSource.loop = false;
+            AudioMusicVoice = new AudioMusic();
+            AudioMusicVoice.MusicSource = gameObjectVoice.GetComponent<AudioSource>();
+            AudioMusicVoice.MusicSource.playOnAwake = false;
+            AudioMusicVoice.MusicSource.loop = false;
 
             GameObject gameObjectBtnClick = new GameObject("BtnClick", typeof(AudioSource));
             gameObjectBtnClick.transform.SetParent(_transform);
-            _audioMusicBtnClick = new AudioMusic();
-            _audioMusicBtnClick.MusicSource = gameObjectBtnClick.GetComponent<AudioSource>();
-            _audioMusicBtnClick.MusicSource.playOnAwake = false;
-            _audioMusicBtnClick.MusicSource.loop = false;
+            AudioMusicBtnClick = new AudioMusic();
+            AudioMusicBtnClick.MusicSource = gameObjectBtnClick.GetComponent<AudioSource>();
+            AudioMusicBtnClick.MusicSource.playOnAwake = false;
+            AudioMusicBtnClick.MusicSource.loop = false;
             
             GameObject gameObjectUISound = new GameObject("UISound", typeof(AudioSource));
             gameObjectUISound.transform.SetParent(_transform);
-            _audioMusicUISound = new AudioMusic();
-            _audioMusicUISound.MusicSource = gameObjectUISound.GetComponent<AudioSource>();
-            _audioMusicUISound.MusicSource.playOnAwake = false;
-            _audioMusicUISound.MusicSource.loop = false;
+            AudioMusicUISound = new AudioMusic();
+            AudioMusicUISound.MusicSource = gameObjectUISound.GetComponent<AudioSource>();
+            AudioMusicUISound.MusicSource.playOnAwake = false;
+            AudioMusicUISound.MusicSource.loop = false;
             
             GameObject gameObjectAudioEffect = new GameObject("AudioEffect", typeof(AudioSource));
             gameObjectAudioEffect.transform.SetParent(_transform);
-            _audioMusicAudioEffect = new AudioMusic();
-            _audioMusicAudioEffect.MusicSource = gameObjectAudioEffect.GetComponent<AudioSource>();
-            _audioMusicAudioEffect.MusicSource.playOnAwake = false;
-            _audioMusicAudioEffect.MusicSource.loop = false;
+            AudioMusicAudioEffect = new AudioMusic();
+            AudioMusicAudioEffect.MusicSource = gameObjectAudioEffect.GetComponent<AudioSource>();
+            AudioMusicAudioEffect.MusicSource.playOnAwake = false;
+            AudioMusicAudioEffect.MusicSource.loop = false;
             
             _audioMusicAudioEffect3D = new AudioEffect();
 
@@ -97,22 +96,18 @@ namespace F8Framework.Core
         /// <param name="audioMixer"></param>
         public void SetAudioMixer(AudioMixer audioMixer)
         {
-            _audioMusic.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/Music")[0];
-            _audioMusicVoice.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/Voice")[0];
-            _audioMusicBtnClick.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
-            _audioMusicUISound.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
-            _audioMusicAudioEffect.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
+            AudioMusic.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/Music")[0];
+            AudioMusicVoice.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/Voice")[0];
+            AudioMusicBtnClick.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
+            AudioMusicUISound.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
+            AudioMusicAudioEffect.MusicSource.outputAudioMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
             _audioEffectMixerGroup = audioMixer.FindMatchingGroups("Master/SoundFx")[0];
             _audioMixer = audioMixer;
         }
         
         public void OnUpdate()
         {
-            _audioMusic.Tick();
-            _audioMusicVoice.Tick();
-            _audioMusicBtnClick.Tick();
-            _audioMusicUISound.Tick();
-            _audioMusicAudioEffect.Tick();
+            
         }
         
         public void OnLateUpdate()
@@ -128,11 +123,11 @@ namespace F8Framework.Core
         public void OnTermination()
         {
             StopAll();
-            Tween.Instance.CancelTween(_audioMusic.AudioTween);
-            Tween.Instance.CancelTween(_audioMusicVoice.AudioTween);
-            Tween.Instance.CancelTween(_audioMusicBtnClick.AudioTween);
-            Tween.Instance.CancelTween(_audioMusicUISound.AudioTween);
-            Tween.Instance.CancelTween(_audioMusicAudioEffect.AudioTween);
+            Tween.Instance.CancelTween(AudioMusic.AudioTween);
+            Tween.Instance.CancelTween(AudioMusicVoice.AudioTween);
+            Tween.Instance.CancelTween(AudioMusicBtnClick.AudioTween);
+            Tween.Instance.CancelTween(AudioMusicUISound.AudioTween);
+            Tween.Instance.CancelTween(AudioMusicAudioEffect.AudioTween);
             
             Destroy(gameObject);
         }
@@ -142,7 +137,7 @@ namespace F8Framework.Core
         // 设置背景音乐播放完成回调
         public void SetMusicComplete(Action callback)
         {
-            _audioMusic.OnComplete = callback;
+            AudioMusic.OnComplete = callback;
         }
         
         /// <summary>
@@ -159,21 +154,21 @@ namespace F8Framework.Core
             {
                 return;
             }
-            if (priority < _audioMusic.Priority)
+            if (priority < AudioMusic.Priority)
             {
                 return;
             }
-            _audioMusic.Load(assetName, callback, loop, priority, fadeDuration);
+            AudioMusic.Load(assetName, callback, loop, priority, fadeDuration);
         }
         
         // 设置背景乐播放进度
         public float SetProgressMusic
         {
-            set => _audioMusic.Progress = value;
+            set => AudioMusic.Progress = value;
         }
         
         // 获取背景音乐播放进度
-        public float ProgressMusic => _audioMusic.Progress;
+        public float ProgressMusic => AudioMusic.Progress;
         
         // 获取和设置背景音乐音量
         public float VolumeMusic
@@ -183,7 +178,7 @@ namespace F8Framework.Core
             {
                 _volumeMusic = value;
                 StorageManager.Instance.SetFloat(_volumeMusicKey, value);
-                _audioMusic.MusicSource.volume = value;
+                AudioMusic.MusicSource.volume = value;
             }
         }
         
@@ -203,7 +198,7 @@ namespace F8Framework.Core
                 StorageManager.Instance.SetBool(_switchMusicKey, value);
                 if (!value)
                 {
-                    _audioMusic.MusicSource.Stop();
+                    AudioMusic.MusicSource.Stop();
                 }
             }
         }
@@ -213,7 +208,7 @@ namespace F8Framework.Core
         // 设置人声播放完成回调
         public void SetVoiceComplete(Action callback)
         {
-            _audioMusicVoice.OnComplete = callback;
+            AudioMusicVoice.OnComplete = callback;
         }
 
         // 播放人声
@@ -223,21 +218,21 @@ namespace F8Framework.Core
             {
                 return;
             }
-            if (priority < _audioMusicVoice.Priority)
+            if (priority < AudioMusicVoice.Priority)
             {
                 return;
             }
-            _audioMusicVoice.Load(assetName, callback, loop, priority, fadeDuration);
+            AudioMusicVoice.Load(assetName, callback, loop, priority, fadeDuration);
         }
         
         // 设置人声播放进度
         public float SetProgressVoice
         {
-            set => _audioMusicVoice.Progress = value;
+            set => AudioMusicVoice.Progress = value;
         }
         
         // 获取人声播放进度
-        public float ProgressVoice => _audioMusicVoice.Progress;
+        public float ProgressVoice => AudioMusicVoice.Progress;
         
         // 获取和设置人声音量
         public float VolumeVoice
@@ -247,7 +242,7 @@ namespace F8Framework.Core
             {
                 _volumeVoice = value;
                 StorageManager.Instance.SetFloat(_volumeVoiceKey, value);
-                _audioMusicVoice.MusicSource.volume = value;
+                AudioMusicVoice.MusicSource.volume = value;
             }
         }
 
@@ -267,7 +262,7 @@ namespace F8Framework.Core
                 StorageManager.Instance.SetBool(_switchVoiceKey, value);
                 if (!value)
                 {
-                    _audioMusicVoice.MusicSource.Stop();
+                    AudioMusicVoice.MusicSource.Stop();
                 }
             }
         }
@@ -282,9 +277,9 @@ namespace F8Framework.Core
             {
                 _volumeAudioEffect = value;
                 StorageManager.Instance.SetFloat(_volumeAudioEffectKey, value);
-                _audioMusicBtnClick.MusicSource.volume = value;
-                _audioMusicUISound.MusicSource.volume = value;
-                _audioMusicAudioEffect.MusicSource.volume = value;
+                AudioMusicBtnClick.MusicSource.volume = value;
+                AudioMusicUISound.MusicSource.volume = value;
+                AudioMusicAudioEffect.MusicSource.volume = value;
             }
         }
 
@@ -304,9 +299,9 @@ namespace F8Framework.Core
                 StorageManager.Instance.SetBool(_switchAudioEffectKey, value);
                 if (!value)
                 {
-                    _audioMusicBtnClick.MusicSource.Stop();
-                    _audioMusicUISound.MusicSource.Stop();
-                    _audioMusicAudioEffect.MusicSource.Stop();
+                    AudioMusicBtnClick.MusicSource.Stop();
+                    AudioMusicUISound.MusicSource.Stop();
+                    AudioMusicAudioEffect.MusicSource.Stop();
                 }
             }
         }
@@ -316,7 +311,7 @@ namespace F8Framework.Core
         // 设置按钮音效播放完成回调
         public void SetBtnClickComplete(Action callback)
         {
-            _audioMusicBtnClick.OnComplete = callback;
+            AudioMusicBtnClick.OnComplete = callback;
         }
 
         // 播放按钮音效
@@ -326,11 +321,11 @@ namespace F8Framework.Core
             {
                 return;
             }
-            if (priority < _audioMusicBtnClick.Priority)
+            if (priority < AudioMusicBtnClick.Priority)
             {
                 return;
             }
-            _audioMusicBtnClick.Load(assetName, callback, loop, priority, fadeDuration);
+            AudioMusicBtnClick.Load(assetName, callback, loop, priority, fadeDuration);
         }
         
         /*----------UI音效特效----------*/
@@ -338,7 +333,7 @@ namespace F8Framework.Core
         // 设置UI音效播放完成回调
         public void SetUISoundComplete(Action callback)
         {
-            _audioMusicUISound.OnComplete = callback;
+            AudioMusicUISound.OnComplete = callback;
         }
 
         // 播放UI音效
@@ -348,11 +343,11 @@ namespace F8Framework.Core
             {
                 return;
             }
-            if (priority < _audioMusicUISound.Priority)
+            if (priority < AudioMusicUISound.Priority)
             {
                 return;
             }
-            _audioMusicUISound.Load(assetName, callback, loop, priority, fadeDuration);
+            AudioMusicUISound.Load(assetName, callback, loop, priority, fadeDuration);
         }
                 
         /*----------音效特效----------*/
@@ -360,7 +355,7 @@ namespace F8Framework.Core
         // 设置音效特效播放完成回调
         public void SetAudioEffectComplete(Action callback)
         {
-            _audioMusicAudioEffect.OnComplete = callback;
+            AudioMusicAudioEffect.OnComplete = callback;
         }
 
         // 播放音效特效
@@ -370,11 +365,11 @@ namespace F8Framework.Core
             {
                 return;
             }
-            if (priority < _audioMusicAudioEffect.Priority)
+            if (priority < AudioMusicAudioEffect.Priority)
             {
                 return;
             }
-            _audioMusicAudioEffect.Load(assetName, callback, loop, priority, fadeDuration);
+            AudioMusicAudioEffect.Load(assetName, callback, loop, priority, fadeDuration);
         }
         
         /*----------一次性3D音效特效----------*/
@@ -403,37 +398,37 @@ namespace F8Framework.Core
         /*----------全局控制----------*/
         public void ResumeAll()
         {
-            _audioMusic.MusicSource.Play();
-            _audioMusicVoice.MusicSource.Play();
-            _audioMusicBtnClick.MusicSource.Play();
-            _audioMusicUISound.MusicSource.Play();
-            _audioMusicAudioEffect.MusicSource.Play();
+            AudioMusic.Resume();
+            AudioMusicVoice.Resume();
+            AudioMusicBtnClick.Resume();
+            AudioMusicUISound.Resume();
+            AudioMusicAudioEffect.Resume();
         }
         
         public void PauseAll() {
-            _audioMusic.MusicSource.Pause();
-            _audioMusicVoice.MusicSource.Pause();
-            _audioMusicBtnClick.MusicSource.Pause();
-            _audioMusicUISound.MusicSource.Pause();
-            _audioMusicAudioEffect.MusicSource.Pause();
+            AudioMusic.Pause();
+            AudioMusicVoice.Pause();
+            AudioMusicBtnClick.Pause();
+            AudioMusicUISound.Pause();
+            AudioMusicAudioEffect.Pause();
         }
         
         public void StopAll()
         {
-            _audioMusic.MusicSource.Stop();
-            _audioMusicVoice.MusicSource.Stop();
-            _audioMusicBtnClick.MusicSource.Stop();
-            _audioMusicUISound.MusicSource.Stop();
-            _audioMusicAudioEffect.MusicSource.Stop();
+            AudioMusic.Stop();
+            AudioMusicVoice.Stop();
+            AudioMusicBtnClick.Stop();
+            AudioMusicUISound.Stop();
+            AudioMusicAudioEffect.Stop();
         }
         
         public void UnloadAll(bool unloadAllLoadedObjects = true)
         {
-            _audioMusic.UnloadAll(unloadAllLoadedObjects);
-            _audioMusicVoice.UnloadAll(unloadAllLoadedObjects);
-            _audioMusicBtnClick.UnloadAll(unloadAllLoadedObjects);
-            _audioMusicUISound.UnloadAll(unloadAllLoadedObjects);
-            _audioMusicAudioEffect.UnloadAll(unloadAllLoadedObjects);
+            AudioMusic.UnloadAll(unloadAllLoadedObjects);
+            AudioMusicVoice.UnloadAll(unloadAllLoadedObjects);
+            AudioMusicBtnClick.UnloadAll(unloadAllLoadedObjects);
+            AudioMusicUISound.UnloadAll(unloadAllLoadedObjects);
+            AudioMusicAudioEffect.UnloadAll(unloadAllLoadedObjects);
             _audioMusicAudioEffect3D.UnloadAll(unloadAllLoadedObjects);
         }